% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqr.R
\name{dq_rpt}
\alias{dq_rpt}
\title{Report DQ}
\usage{
dq_rpt(DF)
}
\arguments{
\item{DF}{A data.frame.}
}
\value{
TBD
}
\description{
\code{dq_rpt(x)} returns the quality data.frame.
}
\examples{
x <- data.frame(a = letters[1:9], b = LETTERS[1:9], c = c(1:9))
dq_rpt(x)
}
