% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqr.R
\docType{package}
\name{dqr-package}
\alias{dqr}
\alias{dqr-package}
\title{dqr: Data Quality for R}
\description{
Data quality checking and reporting.
}
\details{
This package will 
ease the pain of assessing and reporting data quality for a dataframe
or a set of them
}
\author{
\strong{Maintainer}: Mick Mioduszewski \email{mick@mioduszewski.net}

}
\keyword{internal}
